Add soft delete for articles
Add the 'discard' gem to support soft deletion/archiving of articles.

At my current job, soft deletion is implemented in several ways, and
in some cases, not implemented at all. Additionally, we rely on these
various implementations somewhat heavily when updating the status of
real estate transactions, and as a result we get a lot of unexpected
behavior. I'm using this toy app to explore options for a single
implementation of a solution, with the hope that we'll be able to
adopt it in our repo.